+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### 功能构建
 
+- `webapp-pc` 子模块功能优化。
+  - `src/views/nodes/elementPlus/settingrepo/settingNode/subPanels/FileListNodeSubEditPanel.vue` 样式优化。
+
 - `webapp-pc` 子模块文件优化注释、文档注释格式、代码换行格式。
   - src/components/elementPlus\*/*.ts。
   - src/composables/*.ts。